requirements.txt
tensorflow
keras
opencv-python
scikit-learn
pandas
numpy
matplotlib
seaborn
fastapi
uvicorn
python-multipart
pillow
streamlit
requests
